Reverted to previous commits count version
Reverted to the previous commit retrieval function for versioning. The commit limit is now irrelevant because the organization's repository, following the recent fork, will be synchronized with the server. The number of commits for versioning will be determined from this repository, making the previous limit unnecessary, as it is now the responsibility of the school to manage.